2% performance gains by changing Offset type from int32_t to int64_t

+++ src/brainfuck/types.h
@@ -29,7 +29,7 @@ enum Type : uint64_t {
 };
 
 using Value = int64_t;
-using Offset = int32_t;
+using Offset = int64_t;
 
 struct Instruction {
   Type type = NOOP;
